restructure Milsestone components
