+++ assets/app/js/components/Milestone/CompletedRow.jsx
