+++ assets/app/js/containers/ProjectMilestoneView.jsx
@@ -1,13 +1,10 @@
-import React, { Component, PropTypes } from 'react'
-import { connect } from 'react-redux'
-import MilestoneView from '../components/MilestoneView.jsx'
-import {getProjectMilestones, getProjectTasks} from '../selector'
+import { connect } from 'react-redux';
+import MilestoneView from '../components/Milestone/MilestoneView.jsx';
+import { getProjectMilestones, getProjectTasks } from '../selector';
 
-const mapStateToProps = (state, ownProps) => {
-  return {
-    milestones: getProjectMilestones(state),
-    tasks: getProjectTasks(state),
-  }
-}
+const mapStateToProps = (state) => ({
+  milestones: getProjectMilestones(state),
+  tasks: getProjectTasks(state),
+});
 
-export default connect(mapStateToProps)(MilestoneView)
+export default connect(mapStateToProps)(MilestoneView);
